backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
httpx==0.27.0
pydantic==2.8.2
pydantic-settings==2.4.0
pathspec==0.12.1
aiofiles==24.1.0
openai==1.40.0
